requirements.txt
Django~=5.1
pip~=24.2
Pillow~=10.4.0
django-ckeditor~=6.7.1
asgiref~=3.8.1
sqlparse~=0.5.1